requirements.txt
ipinfo
geocoder
phonenumber
bs4
rich
colorama
keyboard
requests
